# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CeetFeedback API",
        "description": "API documentation for CeetFeedback",
        "contact": {
            "email": "blezanatanael@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "V1"
    },
    "paths": {
        "/api/v1/departements": {
            "get": {
                "tags": [
                    "Departements"
                ],
                "summary": "Récupérer la liste des départements",
                "operationId": "fb998789a307fcc64f30715c68292db3",
                "responses": {
                    "200": {
                        "description": "Liste des départements récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DepartementSchema"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "post": {
                "tags": [
                    "Departements"
                ],
                "summary": "Créer un nouveau département",
                "operationId": "c771189d3d67cf3a86c54a554d377859",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DepartementSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Département créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Département créé avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/DepartementSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/v1/departements/{id}": {
            "put": {
                "tags": [
                    "Departements"
                ],
                "summary": "Mettre à jour un département",
                "operationId": "a04511f4377651a6afabb731b7248498",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du département à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DepartementSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Département mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Département mis à jour avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/DepartementSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Département non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Departements"
                ],
                "summary": "Supprimer un département (soft delete)",
                "operationId": "4aa5c1cfee87e5bc025fc28b133b2052",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du département à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Département supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Département supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Département non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/departements/all": {
            "get": {
                "tags": [
                    "Departements"
                ],
                "summary": "Récupérer tous les départements",
                "operationId": "b8e07484ed93d58a3a2b22ace2c64f94",
                "responses": {
                    "200": {
                        "description": "Liste complète des départements",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DepartementSchema"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "DepartementSchema": {
                "title": "Département",
                "description": "Ressource département",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "slug": {
                        "type": "string",
                        "example": "informatique-abc123"
                    },
                    "name": {
                        "type": "string",
                        "example": "Informatique"
                    },
                    "description": {
                        "type": "string",
                        "example": "Département en charge de l'informatique."
                    },
                    "deletedept": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-18T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-18T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}